requirements.txt
Flask==3.0.0
gunicorn==21.2.0
stripe==8.0.0
jsonschema==4.20.0
PyJWT==2.8.0
requests==2.32.4
sentry-sdk[flask]==1.40.0
APScheduler==3.10.4
scikit-learn==1.5.2
notion-client
reportlab
psycopg2-binary
pytest
cloudflare
gunicorn==21.2.0
stripe==8.0.0
